+++ Assets/Bacon/JiaoType.cs
@@ -5,10 +5,10 @@ using System.Text;
 
 namespace Bacon {
     class JiaoType {
-        public static uint NONE = 0;
-        public static uint GANGSHANGHUA = 1;
-        public static uint GANGSHANGPAO = 2;
-        public static uint QIANGANGHU = 3;
-        public static uint ZIMO = 4;
+        public static long NONE = 0;
+        public static long GANGSHANGHUA = 1;
+        public static long GANGSHANGPAO = 2;
+        public static long QIANGANGHU = 3;
+        public static long ZIMO = 4;
     }
 }
