+++ Assets/Bacon/OpCodes.cs
@@ -5,12 +5,12 @@ using System.Text;
 
 namespace Bacon {
     class OpCodes {
-        public static uint OPCODE_NONE = 0;
-        public static uint OPCODE_PENG = 1;
-        public static uint OPCODE_BUGANG = 2;
-        public static uint OPCODE_ZHIGANG = 3;
-        public static uint OPCODE_ANGANG = 4;
-        public static uint OPCODE_HU = 5;
-        public static uint OPCODE_GUO = 6;
+        public static long OPCODE_NONE = 0;
+        public static long OPCODE_PENG = 1;
+        public static long OPCODE_BUGANG = 2;
+        public static long OPCODE_ZHIGANG = 3;
+        public static long OPCODE_ANGANG = 4;
+        public static long OPCODE_HU = 5;
+        public static long OPCODE_GUO = 6;
     }
 }
